% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chow.liu.R
\name{chow.liu}
\alias{chow.liu}
\title{Chow-liu algorithm}
\usage{
chow.liu(S)
}
\arguments{
\item{S}{correlation matrix}
}
\description{
This internal function finds the maximum likelihood tree over a set of observed variables.
It outputs the minimum cost spanning tree on the set of correlations.
}
\examples{


}
\keyword{Chow-Liu}

