% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.deg2.R
\name{contr.deg2}
\alias{contr.deg2}
\title{Contract degree two nodes}
\usage{
contr.deg2(T)
}
\arguments{
\item{T}{a tree}
}
\description{
This is an internal function that is used in the M-step to output a binary tree. It contracts degree two nodes
that do not correspond to observed variables.
}
\keyword{Chow-Liu}
\keyword{Gaussian}
\keyword{algortihm}
\keyword{phylogenetic}
\keyword{trees,}

