% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.step.R
\name{E.step}
\alias{E.step}
\title{E-step}
\usage{
E.step(X, S)
}
\arguments{
\item{X}{the matrix with the observations}

\item{S}{correlation matrix over the full vector including the hidden variables (obtained in the E-step)}
}
\description{
This is the E-step of the structural EM algorithm.
It outputs the minimum cost spanning tree on the set of correlations.
}
\examples{


}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

