% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.bad.nodes.R
\name{expand.bad.nodes}
\alias{expand.bad.nodes}
\title{Expands high degree nodes into many ternary nodes.}
\usage{
expand.bad.nodes(T)
}
\arguments{
\item{T}{A tree}
}
\description{
This function takes all inner nodes of degree more than 4
}
\examples{


}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

