% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.internal.leaves.R
\name{expand.internal.leaves}
\alias{expand.internal.leaves}
\title{Moves internal nodes that represent observed variables to a leaf node.}
\usage{
expand.internal.leaves(T)
}
\arguments{
\item{T}{A tree}
}
\description{
This function moves all observed (green) inner nodes outside.
This is done by replacing any such node with an unobserved (red) copy which is
zero length apart.
}
\examples{


}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

