% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.binary.R
\name{get.binary}
\alias{get.binary}
\title{Get a binary tree}
\usage{
get.binary(T)
}
\arguments{
\item{T}{A tree}
}
\description{
This function takes the tree (of class 'igraph') output by the Chow-Liu algorithm and outputs a binary tree with the
same observed likelihood.
}
\examples{


}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

