% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.corr.R
\name{get.corr}
\alias{get.corr}
\title{Get the correlation matrix}
\usage{
get.corr(T)
}
\arguments{
\item{T}{A tree of class igraph}
}
\description{
This function takes the tree (of class 'igraph') with distances between the tips and outputs
the corresponding correlation matrix
}
\examples{


}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

