% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.corr0.R
\name{get.corr0}
\alias{get.corr0}
\title{Get the correlation matrix}
\usage{
get.corr0(T0)
}
\arguments{
\item{T}{A tree of class phylo}
}
\description{
This function takes a phylogenetic tree (class "phylo")
returns the correlation matrix by taking exp(-dij)
}
\examples{


}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

