% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dist.R
\name{get.dist}
\alias{get.dist}
\title{Get distances}
\usage{
get.dist(R, eps = 1e-10)
}
\arguments{
\item{R}{a correlation matrix}
}
\description{
This function takes a correlation matrix with possibly negative entries, and a threshold eps>0
returns the distance matrix where eps is a replacement for negative values in R
}
\examples{

}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

