% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.binary.R
\name{is.binary}
\alias{is.binary}
\title{Check if a tree is binary}
\usage{
is.binary(T, opt = 0)
}
\arguments{
\item{T}{a tree}

\item{opt}{either 0 or 1, opt=1 allows for degree two nodes which are not obsrved}
}
\description{
This function checks is T is a binary tree
}
\examples{

}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

