% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like.R
\name{like}
\alias{like}
\title{Compute the likelihood function}
\usage{
like(S, X)
}
\arguments{
\item{S}{covariance matrix}

\item{X}{teh data matrix}
}
\description{
This function computes the Gaussian likelihood function
}
\examples{

}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

