% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.step.R
\name{M.step}
\alias{M.step}
\title{The M-step of teh EM algorithm}
\usage{
M.step(S)
}
\arguments{
\item{S}{covariance matrix over the whole vector (both observed and hidden variables)}
}
\description{
This function performs teh M-step of the EM algorithm.
In the M-step we run chow-liu to get a new update of the tree and the big correlation matrix.
The problem is that the tree does not always make sense so we need to replace it with
a binary tree whose leaves correspond exactly to observed nodes. Luckily such a tree, giving
exactly the same observed likelihood, always exists.
}
\examples{

}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

