% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.data.R
\name{normalize.data}
\alias{normalize.data}
\title{The M-step of teh EM algorithm}
\usage{
normalize.data(dat)
}
\arguments{
\item{X}{The data matrix}
}
\description{
This function takes a data matrix and
returns normalized data with zero mean and unit variances.
}
\examples{

}
\keyword{EM}
\keyword{algorithm}
\keyword{structural}

