% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strEM.R
\name{strEM}
\alias{strEM}
\title{Structural EM function}
\usage{
strEM(dat, T0, tol = 1e-07)
}
\arguments{
\item{dat}{The data matrix.}

\item{T0}{the tree of class phylo (together with edge lengths) used as the starting point}

\item{tol}{absolute tolerance for convergence}
}
\description{
This is the main function that performs the structural EM algorithm.
}
\examples{
m <- 5 # number of leaves
N <- 30 # sample size
Ttr <- rtree(m,rooted=FALSE,br=rep(-log(.6),7))
Ttr$tip.label <- 1:m # label leaves by 1:m
Str <- get.corr0(Ttr)
Rtr <- Str[1:m,1:m]
dat <- mvrnorm(N, rep(0,m), Rtr)
D <- get.dist(cor(dat))
T0 <- fastme.bal(D)
T0$edge.length <- (T0$edge.length>0)*T0$edge.length
res <- strEM(dat,T0,tol=1e-6)
}
\keyword{Chow-Liu}
\keyword{Gaussian}
\keyword{algortihm}
\keyword{phylogenetic}
\keyword{trees,}

